requirements.txt
# Data processing
duckdb>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Machine learning
scikit-learn>=1.3.0
scipy

# Visualization
matplotlib>=3.7.0

# Excel output
openpyxl>=3.1.0

# Feature attribution (feature_importance.py only)
shap>=0.44.0
joblib>=1.3.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0


fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9